requirements.txt
pinecone-client
transformers
torch
pandas